Move base store
