+++ mypy.ini
@@ -45,6 +45,7 @@ files =
   synapse/storage/__init__.py,
   synapse/storage/_base.py,
   synapse/storage/background_updates.py,
+  synapse/storage/databases/main/_base.py,
   synapse/storage/databases/main/appservice.py,
   synapse/storage/databases/main/events.py,
   synapse/storage/databases/main/keys.py,
