+++ synapse/storage/databases/main/monthly_active_users.py
@@ -16,8 +16,8 @@ import logging
 from typing import Dict, List
 
 from synapse.metrics.background_process_metrics import wrap_as_background_process
-from synapse.storage._base import SQLBaseStore
 from synapse.storage.database import DatabasePool, make_in_list_sql_clause
+from synapse.storage.databases.main._base import SQLBaseStore
 from synapse.util.caches.descriptors import cached
 
 logger = logging.getLogger(__name__)
